refactor: move InfoModal to global scope
