+++ app/component-library/components-temp/CustomSpendCap/CustomSpendCap.tsx
@@ -3,7 +3,7 @@ import React, { useEffect, useState } from 'react';
 import { Pressable, View } from 'react-native';
 
 import { strings } from '../../../../locales/i18n';
-import InfoModal from '../../../components/UI/Swaps/components/InfoModal';
+import InfoModal from '../../../components/Base/InfoModal';
 import { TOKEN_APPROVAL_SPENDING_CAP } from '../../../constants/urls';
 import formatNumber from '../../../util/formatNumber';
 import { isNumber } from '../../../util/number';
