+++ app/components/UI/Swaps/index.js
@@ -62,7 +62,7 @@ import TokenSelectModal from './components/TokenSelectModal';
 import SlippageModal from './components/SlippageModal';
 import useBalance from './utils/useBalance';
 import useBlockExplorer from './utils/useBlockExplorer';
-import InfoModal from './components/InfoModal';
+import InfoModal from '../../Base/InfoModal';
 import { AlertType } from '../../Base/Alert';
 import { isZero, gte } from '../../../util/lodash';
 import { useTheme } from '../../../util/theme';
