+++ app/components/UI/TimeEstimateInfoModal/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View } from 'react-native';
 import Text from '../../Base/Text';
-import InfoModal from '../Swaps/components/InfoModal';
+import InfoModal from '../../Base/InfoModal';
 import PropTypes from 'prop-types';
 import { strings } from '../../../../locales/i18n';
 import AppConstants from '../../../core/AppConstants';
